Allows user to save filtered images

+++ src/main/java/com/kc1337/repositories/ImageRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by student on 7/10/17.
  */
-public interface ImageRepository extends CrudRepository<Image,Integer> {
+public interface ImageRepository extends CrudRepository<Image,Long> {
 
 }
